Extract `RawBitIter`
